refactor: split DefaultMatcher pipe-modifiers into per-category modules
Follow-up to #1861 / #1894 (the per-modifier category reorganization deferred
from #1862/#1863). The behavior of each pipe modifier was spread across
hand-enumerated `match` arms in `pipe_element.rs` (`pipe_pattern`,
`is_eqfield_match`, `get_eqfield`) and `default_matcher.rs` (`is_match_inner`,
`build_fast_match`), so adding a Sigma modifier meant editing several dispatch
sites.

Move each modifier's logic into a per-category module under
`src/detections/rule/matchers/modifiers/`:

- `string`   - startswith/endswith/contains wrapping + wildcard->regex
- `regex`    - the re|i / re|m / re|s inline-flag prefixes
- `numeric`  - gt/lt/gte/lte comparison
- `cidr`     - cidr matching
- `fieldref` - equalsfield/endswithfield/fieldref*/exists field references
- `encoding` - base64/base64offset + utf16/wide selection (Encoding/Utf16Kind)

The value-matching categories (cidr/numeric/fieldref) sit behind a small
`ValueMatcher` trait, so `is_match_inner` now calls `pipe.value_match(...)`
instead of a 40-line match; `pipe_pattern`/`build_fast_match` delegate to the
`string`/`regex`/`encoding` modules. The `PipeElement` enum and the `MatchPlan`
normalization from #1862 are unchanged.

Pure reorganization: 157 matcher tests + the full lib suite pass, and detection
output is byte-identical to main - the 6 rule-matching outputs on the sample-evtx
corpus (csv/json-timeline/search) plus a byte-identical 2.9 GB csv-timeline over a
2,838-evtx corpus. (The eid-metrics/logon-summary outputs differ only in
HashMap-seed row order, exactly as they do on main.)
